[ci] Cleanup
- Made each workflow's name consistent
- Rename each workflow file with consistent naming scheme
- Promote flow-ci-ghaction to flow-ci

ghstack-source-id: 490b643dcdbb5f6d853815064a1287109a17d259
Pull Request resolved: https://github.com/facebook/react/pull/30037
