+++ .github/workflows/runtime_commit_artifacts.yml
@@ -1,4 +1,4 @@
-name: Commit Artifacts for Meta WWW and fbsource
+name: (Runtime) Commit Artifacts for Meta WWW and fbsource
 
 on:
   push:
