+++ scripts/tasks/flow-ci.js
@@ -14,24 +14,19 @@ process.on('unhandledRejection', err => {
 const runFlow = require('../flow/runFlow');
 const inlinedHostConfigs = require('../shared/inlinedHostConfigs');
 
-// Parallelize tests across multiple tasks.
-const nodeTotal = process.env.CIRCLE_NODE_TOTAL
-  ? parseInt(process.env.CIRCLE_NODE_TOTAL, 10)
-  : 1;
-const nodeIndex = process.env.CIRCLE_NODE_INDEX
-  ? parseInt(process.env.CIRCLE_NODE_INDEX, 10)
-  : 0;
-
-async function checkAll() {
-  for (let i = 0; i < inlinedHostConfigs.length; i++) {
-    if (i % nodeTotal === nodeIndex) {
-      const rendererInfo = inlinedHostConfigs[i];
-      if (rendererInfo.isFlowTyped) {
-        await runFlow(rendererInfo.shortName, ['check']);
-        console.log();
-      }
-    }
+async function check(shortName) {
+  if (shortName == null) {
+    throw new Error('Expected an inlinedHostConfig shortName');
+  }
+  const rendererInfo = inlinedHostConfigs.find(
+    config => config.shortName === shortName
+  );
+  if (rendererInfo == null) {
+    throw new Error(`Could not find inlinedHostConfig for ${shortName}`);
+  }
+  if (rendererInfo.isFlowTyped) {
+    await runFlow(rendererInfo.shortName, ['check']);
   }
 }
 
-checkAll();
+check(process.argv[2]);
